Added header handling in the server and client
The protocol's header consists of 3 parts:
[Length | uint32] [ID | uint32] [Type | int32]

Length - The length of the message that follows the header
ID - The ID of the connected client (user id)
Type - The action the user is trying to do

In this commit we added support for the new protocol in the client.
In the server there is currently only support for recieving messages.